Mapping (#6)
* saml config

* manifest for saml files

* required

* required

* travis added install

* travis added install

* travis added installs

* extras

* extras

* added

* added

* onelogin configs

* onelogin configs

* using ready

* typo

* email sender

* modified config.py

added CORS variables in order to be possible to enable the CORS. also rearanged the config.py variables and thier comments.

* missing import:

in config.py import is needed in order to use flask_babelex func. also this module requires invenio-saml to be installed - added to the install_requires. for now installing it via github.

* install_requires

changed the branch for invenio-saml

* updated setup.py

changed to custom branch of invenio-saml, for testing the changes on those branch.

* Update config.py

In order to register user_profile via login with OAuth client, one must set this value.

* typo

fixed typo in comment.

* removed entityId

removed manully input for service provider, now it will assign automaticly the metadata url to service provider entityID.

* mapping field

a. changed invenio-saml branch, b. changed mapping field names.

* added missing NameIDFormat

* fix missplaced

name and surname was misplaced

* changed branch